+++ app/examples/circleci/push.yml
@@ -1,0 +1,20 @@
+version: '2.1'
+orbs:
+  docker: circleci/docker@2.4.0
+
+jobs:
+  push:
+    docker:
+      - image: circleci/python:3.7
+    steps:
+      - checkout
+      - docker/pull:
+          images: ubuntu:16.04
+      - docker/push:
+          image: ubuntu
+          tag: '16.04'
+
+workflows:
+  push-images:
+    jobs:
+      - push
